+++ demo/vue/vite.config.ts
@@ -3,6 +3,6 @@ import vue from '@vitejs/plugin-vue'
 import { pretextEditorPlugin } from 'pretext-editor/vite'
 
 export default defineConfig({
-  base: '/pretext-editor/',
+  // base: '/pretext-editor/',
   plugins: [vue(), pretextEditorPlugin()],
 })