+++ src/worker-impl.d.ts
@@ -1,0 +1,4 @@
+declare module '#worker-impl' {
+  export function createWorker(): Worker
+  export function createEagerWorker(): Worker | null
+}